Aiming works
Figured out that there was a displacement due to how the unit circle
works. It starts from the 0 point which points right. From my design of
the cannon in unity, the cannon points directly north. This means it
starts at 90 degrees and I had to add a displacement of 90 degrees.
Must modify cannon to optimize code.