Fixed scaling vs time
Problem: First scaling was too fast
Solution: Put scales expectations into local variables, this way they
won’t be affected by frame changes.

Future problems: Cosmetic, There are dots in the game screen that show
which way it is firing, if it is fine leave it, if its not fine…
Solution: Just turn off the sprite renderer when its not needed and
turn it on when it fires.